# Add reactive-js to perf tests comparison