I just released an alpha of reactive-js (https://github.com/bordoley/reactive-js), a new reactive programming library i've been working on that deeply integrates a cooperate multi-tasking scheduler into the core subscriber type. My main goal when starting the project was to update the Observable contract to play nicely with React's new multitasking scheduler, but pleasantly to my surprise I realized that it also resulted in very fast synchronous observables when run on a virtual time scheduler. Results below are from a mid-2014 macbook pro. 

The library API is still a little unstable, so I'm not sure that I'd merge this yet, but I wanted to share the results at least. 

funrecordss-MacBook-Pro:perf funrecords$ npm run start
```
> perf@0.0.0 start /Users/funrecords/Desktop/Code/callbag-basics/perf
> npm-run-all filter-map-reduce merge combine fold dataflow filter-map-fusion


> perf@0.0.0 filter-map-reduce /Users/funrecords/Desktop/Code/callbag-basics/perf
> NODE_ENV=production node ./filter-map-reduce

filter -> map -> reduce 1000000 integers
-----------------------------------------------
cb-basics    34.97 op/s ±  1.27%   (44 samples)
xstream      14.58 op/s ±  0.87%   (68 samples)
most        513.28 op/s ±  0.53%   (81 samples)
reactive-js  261.07 op/s ±  0.62%   (79 samples)
rx 5         35.39 op/s ±  2.01%   (80 samples)
rx 4          2.06 op/s ±  1.00%   (15 samples)
kefir        11.39 op/s ± 10.36%   (58 samples)
bacon         1.67 op/s ±  4.82%   (13 samples)
highland      4.18 op/s ±  2.50%   (24 samples)
lodash       20.99 op/s ±  0.60%   (39 samples)
Array        17.92 op/s ±  2.28%   (48 samples)
-----------------------------------------------

> perf@0.0.0 merge /Users/funrecords/Desktop/Code/callbag-basics/perf
> NODE_ENV=production node ./merge

merge 100000 x 10 streams
-----------------------------------------------
cb-basics    31.18 op/s ±  1.12%   (72 samples)
xstream      22.62 op/s ±  1.66%   (56 samples)
most        205.51 op/s ±  0.97%   (81 samples)
reactive-js  265.12 op/s ±  0.59%   (79 samples)
rx 5         62.68 op/s ±  0.45%   (72 samples)
rx 4          1.67 op/s ±  0.89%   (13 samples)
kefir        12.50 op/s ±  1.21%   (60 samples)
bacon         1.65 op/s ±  4.49%   (13 samples)
lodash       17.66 op/s ±  2.02%   (48 samples)
Array        40.15 op/s ±  1.13%   (53 samples)
-----------------------------------------------

> perf@0.0.0 combine /Users/funrecords/Desktop/Code/callbag-basics/perf
> NODE_ENV=production node ./combine

combine(add3) -> filter 500000 x 3 integers
-----------------------------------------------
cb-basics    18.93 op/s ±  0.76%   (83 samples)
xstream      19.20 op/s ±  0.84%   (62 samples)
most         70.27 op/s ±  0.56%   (79 samples)
reactive-js   40.42 op/s ±  0.56%   (64 samples)
rx 5         32.27 op/s ±  1.82%   (74 samples)
-----------------------------------------------

> perf@0.0.0 fold /Users/funrecords/Desktop/Code/callbag-basics/perf
> NODE_ENV=production node ./fold

scan -> reduce 1000000 integers
-----------------------------------------------
cb-basics    21.34 op/s ±  0.77%   (53 samples)
xstream      13.39 op/s ±  1.10%   (64 samples)
most        331.09 op/s ±  0.35%   (82 samples)
reactive-js   33.62 op/s ±  3.22%   (77 samples)
rx 5         19.90 op/s ±  2.58%   (51 samples)
rx 4          1.99 op/s ±  1.43%   (14 samples)
kefir        14.96 op/s ±  1.12%   (39 samples)
bacon         1.12 op/s ±  9.35%   (10 samples)
highland      4.40 op/s ±  2.07%   (26 samples)
lodash        6.72 op/s ± 10.51%   (19 samples)
Array         6.30 op/s ±  9.56%   (14 samples)
-----------------------------------------------

> perf@0.0.0 dataflow /Users/funrecords/Desktop/Code/callbag-basics/perf
> node ./dataflow

dataflow for 1000000 source events
-----------------------------------------------
cb-basics     6.73 op/s ±  2.48%   (36 samples)
xstream       4.23 op/s ±  1.55%   (25 samples)
most         14.40 op/s ±  1.05%   (67 samples)
rx 5          6.55 op/s ±  0.80%   (35 samples)
rx 4          0.65 op/s ±  1.69%    (8 samples)
bacon         0.77 op/s ±  6.98%    (8 samples)
-----------------------------------------------

> perf@0.0.0 filter-map-fusion /Users/funrecords/Desktop/Code/callbag-basics/perf
> NODE_ENV=production node ./filter-map-fusion

filter -> map -> fusion 1000000 integers
-----------------------------------------------
cb-basics    19.88 op/s ±  0.79%   (35 samples)
xstream      16.06 op/s ±  0.81%   (73 samples)
most        105.15 op/s ±  0.41%   (79 samples)
reactive-js   42.93 op/s ±  3.49%   (68 samples)
rx 5         18.09 op/s ±  0.90%   (81 samples)
rx 4          1.81 op/s ±  0.94%   (13 samples)
kefir         8.17 op/s ±  1.17%   (42 samples)
bacon         1.34 op/s ±  4.55%   (11 samples)
highland      2.39 op/s ±  1.23%   (16 samples)
lodash       13.84 op/s ±  1.20%   (38 samples)
Array        12.89 op/s ±  1.74%   (36 samples)
-----------------------------------------------
funrecordss-MacBook-Pro:perf funrecords$ 
```