@staltz 2019-12-09
@bordoley That's great that you created a library that turns out to have really good performance. But is this callbag-basics repo the place where you should put these benchmark comparisons? My intention with these benchmarks was to compare callbag-basics with other well known stream libraries, not *all* stream libraries. Your library is still somewhat unknown. I suggest copying the benchmark code into your repo and doing the comparison there. Furthermore, this PR removes the flatten and flattenConcurrently benchmarks, which are ones I'd like to keep. I hope you understand why this PR shouldn't be merged.